piperci-flake8-executor/requirements.txt
PyYAML
requests
flake8
minio
pipercilib@git+https://github.com/afcyber-dream/python-piperci.git